// game/types.ts
export type BuildingType =
  | 'elihut'
  | 'chicken_coop'
  | 'farm_plot'
  | 'town_hall'
  | 'gold_mine'
  | 'elixir_collector'
  | 'gold_storage'
  | 'elixir_storage'
  | 'barracks'
  | 'army_camp'
  | 'cannon'
  | 'archer_tower'
  | 'wizard_tower'
  | 'wall';